Fixed bug in roundtripping an empty file
Firstly, in filehandling,  an empty submitted file was not having its filename and mimetype
kept intact

Secondly, jsonutil was adding the doc_id on encoding to dict .
Unfortunately when an item is created, the doc_id is not known at the
point of conversion. Hence I've moved the addition onto the loading
instead. This means that images look like they don't have their doc_ids
in CouchDB because they are assigned on read.