Add the tests I originally wrote, rename the class and fix the bugs introduced when it was copied in place.
I renamed the class from Store to Source. It's definitely a Source (it
doesn't even have a put method) and the docstring describes it as such.

When it was copied in an "import couchdb" was left out. I guess that
means it's never been used. It's in the package though so it needs
testing and to work. As it happens we changed the exception type some
time ago so, even without the missing import, it would have been
incorrect. Of course, if the tests had been added too then we would have
spotted the problem immediately ;-).