Add options to remove rows with no, i.e. deleted, doc from docs_by_XXX methods.
It's not uncommon to use a secondary database, e.g. a full text index,
to find a set of keys and then hit CouchDB for the docs for those keys.
However, the secondary database may well lag the CouchDB a little.

It typically doesn't matter too much if the secondary database hasn't
got the latest data for a doc but it does matter if the doc has been
deleted.

I've been caught out by this a number of times now (and it's mostly
down to "bad" timing) so the docs_by_XXX methods now have a
remove_rows_with_missing_doc arg (default is False) to remove these rows
before they're returned.