Record ctime and mtime timestamps in a 'metadata' block.
The metadata.ctime and metadata.mtime are set to an ISO 8601 format
string when create (ctime and mtime) and updated (mtime only).

The update of the metadata is designed to be non-destructive so other
libraries or applications are free to store additional metadata ...
as long as it's not called ctime and mtime.

In addition, existing docs with no metadata work - the mtime will be
recorded on next save. (The ctime is never set, that's up to the
application).